This changes everything (#105)
Previously the JavaScript client would always connect the websocket at `/live`. The meant the handler for that endpoint wouldn't know which live views were on the page. It also meant the views had to be spawned for the original request from the browser, before the websocket connects. That lead to a whole bunch of complications:

- The server the view was running on and the server the websocket was on wasn't necessary the same (if you're load balancing multiple servers). Thus we had to have the entire pubsub system just to pass messages between the view and the socket.
- It was a security issue that someone could just `curl` the endpoint to spawn a view. Ideally it should only happen when the socket connects.

But I've realized that if the JavaScript client simply calls the same route as the initial browser request then we can use an extractor+response to figure out if the request is for a socket connection or not but always provide the views to axum-live-view in a consistent way. That meant the views life cycle code and the socket handling code now directly know about each other. That meant:

- The entire pubsub system can go away, because there is only one code path. This should make things faster since we don't have to pay the serialization (and network) cost of sending pubsub messages.
- Reconnecting becomes way easier. The browser simply makes an new websocket request if the socket closes. That'll spawn a new view on the server and go through the extract same flow as when connecting in the first place.

It also made a few things more difficult:

- You can't directly on two views on one socket. Instead you have to use `axum_live_view::live_view::combine` which combines multiple views into one, forwards the messages as necessary, and allows you to customize how the views should be rendered when they're combined.
- Passing messages from the outside into a view becomes a little more manual. Now you'll get a `SelfHandle` in `LiveView::mount` which can be used to send messages to the current view. The idea is that in `mount` you spawn a task to receive messages and send those, via the handle, back to the view.

I also took the chance to refactor most of the TypeScript code especially to make passing input and form data much nicer. There is now less manual deserialization.